feat: add user name extraction and time formatting utilities and refactor participant summary service to use them
